Remove -MD from the CFLAGS in Makefile.defs to not clutter root directory.
The CFLAGS added the generation of dependency files for all invocations of CC.
In the case of gneerating the qstring database, this led to superfluous .d files
in the platform root directory. Dependency files (.P) are added to the build target
directory later and are correctly included by py.mk, so there is no need to do
anything else.